improved messaging and cancecllation handling
